+++ packages/mvc/package.json
@@ -37,6 +37,10 @@
       "types": "./dist/observable.d.ts",
       "default": "./dist/observable.js"
     },
+    "./runtime": {
+      "types": "./dist/runtime.d.ts",
+      "default": "./dist/runtime.js"
+    },
     "./jsx-runtime": {
       "types": "./dist/jsx-runtime.d.ts",
       "default": "./dist/jsx-runtime.js"
